requirements.txt
asgiref==3.11.1
attrs==26.1.0
cffi==2.0.0
cryptography==46.0.6
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.29.0
drf-yasg==1.21.8
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.2.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
python-decouple==3.8
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
setuptools==82.0.1
sqlparse==0.5.5
uritemplate==4.2.0
wheel==0.46.3
